Simplify demos
