+++ examples/react-minimal/src/main.tsx
@@ -4,6 +4,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { api } from "../convex/_generated/api";
 import { App } from "./App";
+import "./index.css";
 
 const convex = new ConvexReactClient(import.meta.env.VITE_CONVEX_URL!);
