+++ examples/react-password/src/main.tsx
@@ -2,8 +2,10 @@ import { ConvexAuthProvider } from "@convex-dev/auth/react";
 import { ConvexReactClient } from "convex/react";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 import { api } from "../convex/_generated/api";
 import { App } from "./App";
+import "./index.css";
 
 const convex = new ConvexReactClient(import.meta.env.VITE_CONVEX_URL!);
 
@@ -16,7 +18,9 @@ createRoot(document.getElementById("root")!).render(
         signOut: api.auth.signOut,
       }}
     >
-      <App />
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
     </ConvexAuthProvider>
   </StrictMode>,
 );
